feat: update services and tests controllers
